feat(phone): add loaders to call history/stats, update history and stats style, cleanup
